# node-output/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ScreenSpace",
        "description": "This is a non commercial project for a software engineering degree. It´s a REST API for cinema management and movie ticket booking mobile app.",
        "contact": {
            "email": "contact@screenspace.com"
        },

        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://virtserver.swaggerhub.com/BREUERANDRES/ScreenSpace-API/1.0.0",
            "description": "SwaggerHub API Auto Mocking"
        },
        {
            "url": "http://localhost:3500/api/v1",
            "description": "Developement URL"
        },
        {
            "url": "https://screenspace-api.azurewebsites.net/api/v1",
            "description": "Production URL"
        }
    ],
...
    "paths": {
...
        "/owners/forgotPassword": {
            "post": {
                "tags": ["Owners"],
                "summary": "Reset password",
                "description": "Sends email with a provisional password and a token",
                "operationId": "forgotPassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/owners_forgotPassword_body"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_7"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Input Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_404"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/owners/resetPassword/{token}": {
            "patch": {
                "tags": ["Owners"],
                "summary": "Reset password",
                "description": "Verifies that token was not modified and sets new password",
                "operationId": "resetPassword",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Hashed token.",
                        "required": true,
                        "style": "simple",
                        "explode": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/resetPassword_token_body"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_8"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Input Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_400_3"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "inline_response_404": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "fail"
                    },
                    "message": {
                        "type": "string",
                        "example": "There is no owner with thar email"
                    }
                }
            },
...
            "owners_forgotPassword_body": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@mail.com"
                    }
                }
            },
            "inline_response_200_7": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Password recovery email sent!"
                    }
                }
            },
            "resetPassword_token_body": {
                "type": "object",
                "properties": {
                    "password": {
                        "type": "string",
                        "example": "newPassword1234"
                    }
                }
            },
            "inline_response_200_8": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Password Updated"
                    }
                }
            },
            "inline_response_400_3": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "fail"
                    },
                    "message": {
                        "type": "string",
                        "example": "Token expired or invalid"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}